Divide server logic into seperate modules
