+++ server.js
@@ -2,14 +2,14 @@
 // Import built-in http built in method
 import express from 'express';
 import cors from 'cors';
-import fs from 'fs';
 import dotenv from 'dotenv';
 import process from 'process';
 import path from 'path';
 import { fileURLToPath } from 'url';
 import { v4 as uuidv4 } from 'uuid';
-import jwt_encode from "jwt-encode";
-import fetch from 'node-fetch';
+import fetchCurrencyData from './services/currencyService.js';
+import { SPARCO_PUB_KEY, SPARCO_SEC_KEY, sparcoPayment } from './services/sparcoService.js';
+import addUserToDataBase from './services/storage.js';
 
 
 // Create an expressJs instance
@@ -17,19 +17,15 @@ const app = express();
 
 // Use the cors middleware 
 app.use(cors({
-  origin: '*',                   // Specify the allowed origin for your frontend
+  origin: '*',                                       // Specify the allowed origin for your frontend
   methods: '*',                                      // allowed HTTP methods
   allowedHeaders: '*',                               // allowed headers
 }));
 
 // Configure and retrieve environment variables
 dotenv.config();
-const SPARCO_PUB_KEY =  process.env.SPARCO_PUB_KEY;
-const SPARCO_SEC_KEY =  process.env.SPARCO_SEC_KEY;
 const payPalId = process.env.PAYPAL_CLIENT_ID;
 
-if (!SPARCO_PUB_KEY) { console.log('Sparco Public Key missing') }
-if (!SPARCO_SEC_KEY) { console.log('Sparco Secrete Key missing') }
 if (!payPalId) { console.log('Paypal client ID missing') }
 
 // Declare port and url variable
@@ -41,13 +37,13 @@ app.use(express.json());
 
 // Serve static files from the 'dist' directory generated by Vite.
 const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename); // Get the current directory
+const __dirname = path.dirname(__filename);
 
 app.use(express.static(path.join(__dirname, 'dist')));
 
 // Listen for requests
 app.listen(port, () => {
-  console.log(`Server is running on https://localhost:${port}`);
+  console.log(`Server is running on http://localhost:${port}`);
 })
 
 // All routes will be handled by serving the 'index.html' file.
@@ -56,37 +52,6 @@ app.get('/', (req, res) => {
 });
 
 // Currency conversion
-const fetchCurrencyData = async (country) => {
-  console.log('Fetching converted currency');
-  let countryCurrency = '', currencyToUsdRate, status;
-  await fetch(`https://restcountries.com/v3.1/name/${country}`)
-  .then((res) => res.json())
-  .then((json) => {
-    countryCurrency = Object.keys(json[0].currencies)[0]
-  })
-  .catch(error => {                                                      // Handle any errors
-    console.error(error);
-    status = false;
-  });
-
-
-  const tempAccessKey = "397012be4f747e005d730440024aaf2a"
-  // Free currency api that only provides the base currency in euros.
-  await fetch(`http://data.fixer.io/api/latest?access_key=${tempAccessKey}`)
-  .then((res) => res.json())
-  .then((json) => {
-    const eurToUsd = 1/json.rates.USD;  // Converte Euro Amount to USD. 1 USD = 1 / 1.20 EUR
-    currencyToUsdRate = json.rates[countryCurrency] * eurToUsd;
-    status = true;
-  })
-  .catch(error => {                                                      // Handle any errors
-    console.error(error);
-    status = false;
-  });
-  return {countryCurrency, currencyToUsdRate, status};
-}
-
-
 app.post(`${baseUrl}/currency-conversion`, async (req, res) => {
   const country = req.body.country;
   const countryCurrencyConversion = await fetchCurrencyData(country);
@@ -103,87 +68,18 @@ app.get(`${baseUrl}/paypal-id`, (req, res) => {
 });
 
 
-// Check sparco payment status
-const sparcoPaymentStatus = (requestRef, encoded_payload) => {
-  // Return a promise
-  return new Promise((resolve) => {
-    // Return failed status if payment is still pending after 3 minutes
-    const timer = setTimeout(() => {
-      console.log("Sparco payment request timed out")
-      clearInterval(interval);
-      resolve('FAILED');
-    }, 180000);
-
-    // Set interval to repeatedly check payment status every 5 seconds
-    const interval = setInterval(async () => {
-      try {
-        const paymentStatus = await fetch(`https://live.sparco.io/gateway/api/v1/transaction/query?reference=${requestRef}`, {
-          headers: { 'token': encoded_payload },
-        });
-        if (paymentStatus.ok) {
-          const response = await paymentStatus.json();
-          console.log(response.status)
-          switch (response.status) {
-            case "TXN_AUTH_UNSUCCESSFUL":
-              console.log(resolve(response.status))
-              clearInterval(interval);
-              clearTimeout(timer);
-              resolve(response.status);         // Resolve the promise with the final status
-              break;
-            case "TXN_FAILED":
-              console.log(resolve(response.status))
-              clearInterval(interval);
-              clearTimeout(timer);
-              resolve(response.status);         // Resolve the promise with the final status
-              break;
-            case "TXN_SUCCESSFUL":
-              clearInterval(interval);
-              clearTimeout(timer);
-              resolve("SUCCESS");               // Resolve the promise with the final status
-              break;
-            default:
-              break;
-          }
-        } else {
-          // Handle non-successful response
-          resolve(paymentStatus.status);       // Resolve the promise with the final status
-          clearInterval(interval);
-          clearTimeout(timer);
-        }
-      } catch(error) {
-        console.log('error posting to sparco', error);
-        clearInterval(interval);
-        clearTimeout(timer);
-        resolve("FAILED");                    // Resolve the promise with the final status
-      }
-    }, 5000);
-  });
-};
-
 // Handle post request from front end
 app.post(baseUrl, async (req, res) => {
   if (!SPARCO_PUB_KEY || !SPARCO_SEC_KEY || !payPalId) {
     res.status(500).send('Internal Server Error(missing server side credentials)');
   }
 
-  const data = req.body;                      // Decoding encrypted data from the front-end
-  
-  // Create simple database
-  const addUserToDataBase = () => {
-    if (!fs.existsSync('./database')){        // Check if /database does not exit
-      fs.mkdir('./database', (err) => {       // Create directory
-        if(err){ console.log(err) }           // Check if there is an error
-      })
-    }
-    fs.writeFile('./database/donors.txt', `\nNAME: ${data.first} ${data.lastName} EMAIL: ${data.email} AMOUNT: $${data.usdAmount}\n`, { flag: 'a' }, (err) => {
-      if (err) { console.error('Error appending data to the file:', err);
-      } else { console.log('Record Saved')}
-    });
-  }
+  const data = req.body;                      // Destructuring data from the front-end
 
   // Handle Mobile Money payement request
   if (data.paymentmode !== 'visa' && SPARCO_PUB_KEY && SPARCO_SEC_KEY) {
     console.log("Payment Method: Mobile Money")
+
     // Create SparcoPay payload
     const payload = {
       customerEmail: data.email,
@@ -199,41 +95,12 @@ app.post(baseUrl, async (req, res) => {
       chargeMe: false,
     };
 
-    // Encode SparcoPay data using Sparco secrete key.
-    const encoded_payload = jwt_encode(payload, SPARCO_SEC_KEY);
+    // Make sparco payment
+    sparcoPayment(payload, res);
 
-    const baseUrl = "https://live.sparco.io/gateway/api/v1/momo/debit";
-
-    // Post data to Sparco API
-    try {
-      const sparcoPaymentRequest = await fetch(baseUrl, {
-        method: 'post',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({"payload": encoded_payload}),
-      });
-
-      if (sparcoPaymentRequest.ok) {
-        console.log('Request successfully sent to Sparco API')
-        const response = await sparcoPaymentRequest.json();                                   // Parse the response data as JSON
-        const sparcoSatus = await sparcoPaymentStatus(response.reference, encoded_payload);   // Passing reference to check payment status
-        if (sparcoSatus == "SUCCESS") {
-          addUserToDataBase()                                      // Include name and amout to database
-          res.status(200).send('successful');                      // Send a successful response.
-        } else {
-          console.log("Sparco fail: ", sparcoSatus)
-          res.status(500).send(sparcoSatus);                       // Send a unsuccessful response.
-        }
-      } else {
-        // Handle non-successful response
-        throw new Error(`HTTP error! Status: ${sparcoPaymentRequest.status}`);
-      }
-    } catch(error) {
-      console.log('error posting to sparco', error);
-      res.status(500).send('Internal Server Error');                  // Return error status to frontend
-    }
   } else if (data.paymentmode === 'visa'){                            // Handle paypal payment status
     if(data.payPalStatus == true){
-      addUserToDataBase()                                             // Include name and amout to database
+      addUserToDataBase(data)                                             // Include name and amout to database
     } else {
       res.status(500).send('Internal Server Error');
     }
